build/image: build the storage-operator image
Add a new image to be built: the container image for the Storage
operator.
Since it's very similar to a LocalImage, except for the building part
(we use the operator-sdk command instead of Docker directly, because we
need to compile the Go code), we introduce a new class (OperatorImage)
that inherits from LocalImage and overload the `_do_build` method.

Note that we take care of enabling Go module and compiling from the
`storage-operator` directory.

Refs: #1410
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>