ci/worker: pre-download Go dependencies
We need to download a fairly big amount of data when building operator
using the Operator SDK (~3G of source repositories).

In order to avoid stressing the CI (and wasting bandwidth) by repeatedly
doing this at build time for every single build, let's pre-download all
that stuff and keep it in the container image.

Note that since we need to copy the `go.sum` and `go.mod` files, we need
to set the build context in the storage-operator directory.

The `pod-linter` is also affected as we use the `operator-sdk` command
during our linting phase.

Closes: #1410
Signed-off-by: Sylvain Laperche <sylvain.laperche@scality.com>