contrib/tf: Split networks and security groups
For better readability, we start re-organizing the files in our root
module. In this commit, we split the `network` definition into one
section dedicated to networks, subnets, etc. and another dedicated to
security groups.
We also extract user-facing variables into a single `variables.tf` file,
which helps readers find out all customization entrypoints for this
module in a single file.