contrib/tf: Split nodes deployment
We split the deployment of Bastion and Bootstrap from `nodes.tf` into
their respective files. Variables and outputs are extracted as well.

Note that the configuration of OpenStack flavours is extended to allow
finer control between the different kinds of nodes (we also increase the
default VM size for Bootstrap and Nodes to reduce the chance of
congestion).