scripts: Ensure archive is mounted before bootstrap
In cases where the initial archive is not mounted at
/srv/scality/metalk8s-<version>, bootstrap will fail to install packages
due to files not being found at the expected path. The required state to
mount archives at their expected path is only called after, in
`metalk8s.roles.bootstrap`.
We add a simple check in the bootstrap script to execute
`metalk8s.archives.mounted` if needed.