contrib/terraform: Add support for RHEL
In 2.4, we have an Eve stage for testing a simple installation on RHEL
7. This used to have its own Terraform definition, but it makes sense to
simply add support for this slight difference in the shared Terraform
module.

This implies specific provisioners to manage RHSM, and a new variable to
choose which OS to user for worker machines (Bastion will remain on
CentOS, for simplicity).